+++ src/main/java/domain/chessPiece/Cannon.java
@@ -1,10 +1,17 @@
-package domain;
+package domain.chessPiece;
+
+import domain.direction.Direction;
+import domain.path.Path;
+import domain.position.ChessPiecePositions;
+import domain.position.ChessPosition;
+import domain.type.ChessPieceType;
+import domain.type.ChessTeam;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static domain.Direction.*;
-import static domain.Direction.RIGHT;
+import static domain.direction.Direction.*;
+import static domain.direction.Direction.RIGHT;
 
 public class Cannon extends UnlimitedMoveChessPiece {
     private static final List<Direction> directions = List.of(UP, DOWN, LEFT, RIGHT);
