+++ src/main/java/domain/chessPiece/LimitedMoveChessPiece.java
@@ -1,4 +1,11 @@
-package domain;
+package domain.chessPiece;
+
+import domain.direction.Directions;
+import domain.path.Path;
+import domain.position.ChessPiecePositions;
+import domain.position.ChessPosition;
+import domain.type.ChessPieceType;
+import domain.type.ChessTeam;
 
 import java.util.ArrayList;
 import java.util.List;
