+++ src/main/java/domain/type/ChessPieceType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.type;
 
 public enum ChessPieceType {
