+++ src/main/java/domain/type/ChessTeam.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.type;
 
 public enum ChessTeam {
