+++ src/test/java/domain/BoardTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.position.ChessPiecePositions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
