+++ src/test/java/domain/CannonTest.java
@@ -4,6 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.Map;
+
+import domain.chessPiece.Cannon;
+import domain.chessPiece.ChessPiece;
+import domain.chessPiece.Pawn;
+import domain.position.ChessPiecePositions;
+import domain.position.ChessPosition;
+import domain.type.ChessTeam;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
