+++ src/test/java/domain/ChessPositionTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.position.ChessPosition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
